requirements.txt
google-cloud-container==2.53.0
google-cloud-compute==1.20.1
jproperties==2.1.2